requirements.txt
python-dotenv>=1.0.0
pymupdf>=1.23.0
sentence-transformers>=2.2.2
faiss-cpu>=1.7.4
huggingface-hub>=0.36.0
tqdm>=4.66.0
numpy>=1.24.0,<2.0.0
